analisis_bibliometrico/backend/requirements.txt
selenium
webdriver-manager
fastapi
uvicorn[standard]
python-multipart
jinja2
bibtexparser
scikit-learn
scipy
sentence-transformers
torch
matplotlib
pandas
wordcloud
plotly
pycountry
fpdf2
kaleido
networkx
rapidfuzz
numpy